requirements.txt
streamlit>=1.28.0
pdfplumber>=0.9.0
openpyxl>=3.0.0
Pillow>=9.0.0
pytesseract
pdf2image